Downloads/Agro-Climate/pyproject.toml
[project]
name = "agro-climate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.42.0",
]
